Fixed implementation and unit test of vector_binary_search_range and implemented ack handling
Conflicts:
	broker/src/subscription.c
	test/sdk/col_vec_test.c

+++ sdk/include/dslink/col/vector.h
@@ -104,7 +104,7 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @param data The data to find
     /// @param cmp_fn The compare function to use
     /// @return The index (>= 0) of the value if found, -1 otherwise
-    long vector_upper_bound(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
+    uint32_t vector_upper_bound(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
 
     /// Searches a range of vector for the data using the given comparison function and a binary search algorithm. 
     /// The vector has to be sorted in order to work.
@@ -114,8 +114,8 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @param lower The lowest index of the range to search in
     /// @param upper The upper (excluded) index of the range to search in
     /// @return The index in the range [lower,upper) of the value if found, -1 otherwise
-    long vector_upper_bound_range(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn, 
-				 uint32_t lower, uint32_t upper);
+    uint32_t vector_upper_bound_range(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn, 
+				      uint32_t lower, uint32_t upper);
 
     /// Frees the internally allocated memory of the vector. Does not free the memory pointed to by the elements.
     /// @param vec The vector
